Initial Commit
The project doesn't do much yet, but it's a good starting point.
We have a dummy daemon (main.rs) that listens on a port and prints out the received messages.
We have a k8s admission webhook that injects an initContainer into created Pods, and the initContainer sends a message to the daemon.